+++ app/Http/Controllers/PaymentController.php
@@ -10,7 +10,7 @@ class PaymentController extends Controller
     public function show(Order $order)
     {
         // Pastikan user hanya bisa melihat order miliknya dan statusnya pending
-        if ($order->user_id !== auth()->id() || $order->status !== 'pending') {
+        if ((int)$order->user_id !== (int)auth()->id() || $order->status !== 'pending') {
             abort(403);
         }
         return view('payment', compact('order'));
